+++ CHANGELOG.md
@@ -5,5 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- Improve DirView cover loading performance
+- Improve video cover extraction performance (reuses output)
+- Implement File.setWallpaper action
+### Fix
+- Image loading blocks ui sometimes
+- GridView search style not discarded on cancel
+
+## [v0.7.0]
 ### Added
 - Initial release features